ART: incorrect locals usage leads to memory issues
BZ: 199154

Implementation of several pre-JNI intrinsics uses a construction, which
is not safe in terms of memory freeing and internal structures usage
after that. In case of this bug, destructor is executed before we use the
object's internal structures. This is a typical issue, which should be fixed.

Change-Id: I500f0051e36854d558be8ea34e70752b2b5647ca
Category: device enablement
Domain: AOSP.ART-Runtime
Origin: internal
Upstream-Candidate: no, need rework
Signed-off-by: avignate <aleksey.v.ignatenko@intel.com>