ART: LSRA changes: reg_storage.h
LSRA (Linear Scan Register Allocation) needs to store more information
in a RegStorage instance.  In particular, both the low and high register
need to be at least 16 bits.  In this case, I am using 32 bits to allow
ignoring overflow.

This change puts the enhancements behind an #define, as increasing the
size from 2 to 12 bytes will affect the performance of the compiler,
even if LSRA is not enabled.  Using a #define will allow measurement of
the cost, as well as ease maintenance if the file changes in AOSP.

Change-Id: Ia328caab62096fe879523d52dee791caedcaa792
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-22736
Category: device enablement
Domain: AOSP.ART-Compiler-BE
Origin: internal
Upstream-Candidate: no, need rework
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>
Reviewed-on: https://android.intel.com:443/332005