Quick compiler: Fix ambiguous LoadValue()
Internal b/17790197 & hat tip to Stephen Kyle

The following custom-edited dex program demonstrated
incorrect code generation caused by type confusion.
In the example, the constant held in v0 is used in both
float and int contexts, and the register class gets
confused at the if-eq.

.method private static getInt()I
    .registers 4
    const/16 v0, 100
    const/4 v1, 1
    const/4 v2, 7
    :loop
    if-eq v2, v0, :done
    add-int v2, v2, v1
    goto :loop
    :done
    add-float v3, v0, v1
    return v2
.end method

The bug was introduced in c/96499, "Quick compiler: reference cleanup"
That CL created a convenience variant of LoadValue which selected the
target register type based on the type of the RegLocation.  It should
not have done so.  The type of a RegLocation is the compiler's best
guess of the Dalvik type - and Dalvik allows constants to be used
in multiple type contexts.  All code generation utilities must specify
desired register class based on the capabilities of the instructions
to be emitted.  In the failing case, OpCmpImmBranch (and
GenCompareZeroAndBranch) will be using core registers, so the
LoadValue must specify  either kCoreReg or kRefReg.

The CL deletes the dangerous LoadValue() variant.

JIRA: https://jira01.devtools.intel.com/browse/GMINL-6172

Category: device enablement
Domain: AOSP.ART-Quick
Upstream-Candidate: no
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6172
Origin: https://android-review.googlesource.com/#/c/109814/
Depends-Change-Id: I7f6813553548b48fb5f54640506d1a361d9db3c1

Change-Id: Ie4ec6e51b19676dbbb9628c72c8b3473a419e7ec