ART: copy dalvikInsn back on throw instruction processing
The potentially throwing instructions in ART are splitted into
CHECK and work portions. On the Mir-To-Lir phase ART copies the
instruction info from the work part to CHECK and uses the CHECK
part for LIR emitting. But in some cases dalvik registers of the
work part might be changed and it may lead to incorrect
calculations.

This patch adds copying of the whole dalvikInsn structure, instead
of the operand field.

Change-Id: I17466169ec86e74f9f202785f1c57c0de9e56d06
Signed-off-by: Alexei Zavjalov <alexei.zavjalov@intel.com>