X86_64 QBE: use RIP addressing
Take advantage of RIP addressing in 64 bit mode to improve the code
generation for accesses to the constant area as well as packed switches.
Avoid computing the address of the start of the method, which is needed
in 32 bit mode.

Updated to match AOSP version already merged.

Change-Id: Ia28c93f98b09939806d91ff0bd7392e58996d108
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-22736
Category: Device Enablement
Domain: AOSP.ART-Quick
Origin: Internal
Upstream-Candidate: yes
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>
Reviewed-on: https://android.intel.com:443/332009