ART: patchoat gets lzma support for compressed odex
To support precompiled compressed odex files the installd needs
the patchoat to support 2 additional options:
--input-oat-lzma-fd and --swap-fd.

The patchoat is capable to inflate the compressed odex file
(specified with either --input-oat-lzma-fd or --input-oat-lzma-file)
into a temporary file (specified with either --swap-fd or --swap-file).
Then the temporary file is patched as if it is specified with
--input-oat-fd or --input-oat-file.

The decompression routines are based on the sources of the lzma tool:
    external/lzma/C/Util/Lzma/LzmaUtil.c

Could not make in-memory decompression and patching because the ElfFile
class maps Elf file using its file descriptor. So a temporary file is
used as it is done by dex2oat.

lzma is linked into patchoat as a static library.

Change-Id: I0a2ee40d6444fed5c06fa6f579e1a8ec250bbbc0
Depends-Change-Id: I1b2d86dd4272389e109f49220b09f9ce4952917f
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-23328
Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Reviewed-on: https://android.intel.com:443/339860