[PORT] ART: Promoted register may be wrong after the load of zero fp constant
Because of Dalvik byte code's lack of immediate typing, the same vreg may be used in core,
ref and fp operations. To properly support GC, we must ensure that if a constant zero is
loaded into a fp view of a vreg, it must also be loaded into the core/ref view. The code
that was responsible for ensuring this failed to handle the case of a vreg having a
non-promoted fp view and a promoted core/ref view.

Change-Id: I56365a676281c2b640c71e26eaf17d72e4ee8d2c
Category: device enablement
Domain: AOSP.ART-Quick
Origin: https://android-review.googlesource.com/#/c/132730/
Upstream-Candidate: no, from aosp
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-17762
Signed-off-by: Vladimir <vladimir.a.ivanov@intel.com>
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>
Reviewed-on: https://android.intel.com:443/340742