ART: patchoat gets zip support for compressed odex
Changed lzma support to zip support due to possible licensing issues.

To support precompiled compressed odex files the installd needs
the patchoat to support 2 additional options: --input-oat-zip-fd
and --swap-fd.

The patchoat is capable to inflate the compressed odex file
(specified with either --input-oat-zip-fd or --input-oat-zip-file)
into a temporary file (specified with either --swap-fd or --swap-file).
Then the temporary file is patched as if it is specified with
--input-oat-fd or --input-oat-file.

Could not make in-memory decompression and patching because the ElfFile
class maps Elf file using its file descriptor. So a temporary file is
used as it is done by dex2oat.

TODO: ART did not have classes for gzip decompression but had zip ones
(ZipArchive and ZipEntry). So the odex files are compressed in zip
files and the only zip entry name is 'odex'. In future it make sense to
change the compression to gzip format.

Change-Id: I9b549a51af60e0ada8ffd9d1989564d0dc8ad9d3
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-26089
Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Reviewed-on: https://android.intel.com:443/341804