Fix race with Heap::ClampGrowthLimit and GC
Added logic for handling the temp bitmap if the GC is running and
the live bitmap is clamped to the mark bitmap. This fixes the SIGABRT
from ClampGrowthLimit if the GC clamped the bitmaps at this point.

Also added locking of the heap_bitmap_lock_ so that added a lock so
that the temp bitmap doesn't change from underneath us.

We cherry-picked this patch from
https://android-review.googlesource.com/#/c/145038/

Change-Id: I6d68e9ead4cb7091472fb3c9ba79c48d63c6d9bc
Category: device enablement
Domain: AOSP.ART-VM
Upstream-Candidate: no, from AOSP
Origin: upstream <https://android-review.googlesource.com/#/c/145038>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-26079
Signed-off-by: Lin Zang <lin.zang@intel.com>
Signed-off-by: Vladimir <vladimir.a.ivanov@intel.com>
Reviewed-on: https://android.intel.com:443/357144